Fixed the bug of cron task deleting job

+++ internal/computing/cron_task.go
@@ -130,7 +130,7 @@ func (task *CronTask) cleanImageResource() {
 
 func (task *CronTask) watchExpiredTask() {
 	c := cron.New(cron.WithSeconds())
-	c.AddFunc("* 0/10 * * * ?", func() {
+	c.AddFunc("* 0/20 * * * ?", func() {
 		defer func() {
 			if err := recover(); err != nil {
 				logs.GetLogger().Errorf("watchExpiredTask catch panic error: %+v", err)
@@ -158,8 +158,12 @@ func (task *CronTask) watchExpiredTask() {
 
 		if len(deployOnK8s) == 0 && len(jobList) > 0 {
 			for _, job := range jobList {
-				if err = NewJobService().DeleteJobEntityBySpaceUuId(job.SpaceUuid, models.JOB_COMPLETED_STATUS); err != nil {
-					logs.GetLogger().Infof("failed to delete job from db, space_uuid: %s, error: %v", job.SpaceUuid, err)
+				if job.CreateTime > 0 {
+					if time.Now().Sub(time.Unix(job.CreateTime, 0)) > time.Hour {
+						if err = NewJobService().DeleteJobEntityBySpaceUuId(job.SpaceUuid, models.JOB_COMPLETED_STATUS); err != nil {
+							logs.GetLogger().Infof("failed to delete job from db, space_uuid: %s, error: %v", job.SpaceUuid, err)
+						}
+					}
 				}
 			}
 		} else if len(deployOnK8s) > 0 && len(jobList) == 0 {
@@ -173,11 +177,6 @@ func (task *CronTask) watchExpiredTask() {
 			for _, job := range jobList {
 				if _, ok := deployOnK8s[job.K8sDeployName]; ok {
 					delete(deployOnK8s, job.K8sDeployName)
-				}
-
-				if time.Now().Sub(time.Unix(job.CreateTime, 0)).Hours() > 2 {
-					deleteSpaceIds = append(deleteSpaceIds, job.SpaceUuid)
-					continue
 				}
 
 				checkFcpJobInfoInChain(job)
